Fix zoom
Key event listener interprets Ctrl- and Ctrl+ as zoom-out and zoom-in
only as by default, and do not try to input '-' or '=' operators.

Previously, Ctrl- resulted in both a zoom-out and an inputted '-'
operator, whereas Ctrl+ did not result in a zoom-in at all.
Checks added so only if the Ctrl key is not pressed is an operator input
called.